llm-agent-bridge/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# gRPC and Protocol Buffers
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.25.0
grpcio-status==1.59.0

# HTTP Client for SDK
httpx==0.25.2
requests==2.31.0

# WebSocket support
websockets==12.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Configuration and Environment
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Async utilities
asyncio-mqtt==0.16.1
aiofiles==23.2.1

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8 